feat: Refactor prelude-generator to use a category-theory-based pipeline
This commit refactors the `prelude-generator` to use a category-theory-based pipeline for processing `use` statements. The new pipeline is more modular, extensible, and easier to reason about.

The pipeline is modeled as a series of functors that map between different categories of code representation. Each functor is a self-contained unit of logic that can be tested independently, and new stages can be added to the pipeline by simply creating new functors and composing them.

This change also fixes a bug where the `toml` crate was unable to serialize the `PathBuf` type, which was causing the pipeline to fail.

Finally, this commit adds documentation for the new pipeline to `lattice.md`.